+++ Procfile
@@ -1,1 +1,2 @@
-web: gunicorn -b 0.0.0.0:$PORT --timeout 300 app:app
+web: gunicorn -b 0.0.0.0:$PORT app:app
+worker: python worker.py
