+++ requirements.txt
@@ -7,3 +7,5 @@ SQLAlchemy==1.3.3
 GeoAlchemy2==0.6.1
 geomet==0.2.0.post2
 pandas==0.24.2
+redis==3.2.1
+rq==1.0
